django_app/django_se4i/requirements.txt
Django
python-dotenv
mysqlclient
black
pylint-django